+++ Bearlike Project/Assets/01.Scripts/Weapon/Gun/GunBase.cs
@@ -138,7 +138,7 @@ namespace Script.Weapon.Gun
             float damageMultiplier = enemyHitbox is TestBodyHitbox bodyHitbox ? bodyHitbox.damageMultiplier : 1f;
             
             // 총의 공격력을 여기서 추가를 할지 아님 state에서 추가를 할지 고민해보자.
-            enemyState.ApplyDamageRPC((status.attack.Current + attack.Current) * damageMultiplier, (CrowdControl)status.property);
+            enemyState.ApplyDamageRPC((int)((status.attack.Current + attack.Current) * damageMultiplier), (CrowdControl)status.property);
         }
 
         #region Bullet Funtion
